pyproject.toml
[tool.poetry]
name = "fastapi-async-benchmark"
version = "0.1.0"
description = "Simple project that evaluates the performance of FastAPI using sync and async operations"
authors = ["Facu Palavecino <facundo.palavecino@muttdata.ai>"]
readme = "README.md"
packages = [{include = "benchmark"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
opentelemetry-distro = "^0.41b0"
opentelemetry-exporter-otlp = "^1.20.0"
opentelemetry-instrumentation-aws-lambda = "0.41b0"
opentelemetry-instrumentation-dbapi = "0.41b0"
opentelemetry-instrumentation-logging = "0.41b0"
opentelemetry-instrumentation-sqlite3 = "0.41b0"
opentelemetry-instrumentation-urllib = "0.41b0"
opentelemetry-instrumentation-wsgi = "0.41b0"
opentelemetry-instrumentation-boto3sqs = "0.41b0"
opentelemetry-instrumentation-botocore = "0.41b0"
opentelemetry-instrumentation-fastapi = "0.41b0"
opentelemetry-instrumentation-grpc = "0.41b0"
opentelemetry-instrumentation-httpx = "0.41b0"
opentelemetry-instrumentation-requests = "0.41b0"
opentelemetry-instrumentation-system-metrics = "0.41b0"
opentelemetry-instrumentation-tortoiseorm = "0.41b0"
opentelemetry-instrumentation-urllib3 = "0.41b0"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
httpx = "^0.25.0"
ruff = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
